requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary>=2.9.0
requests==2.32.4
python-decouple==3.8

# JWT pour Supabase (OBLIGATOIRE)
pyjwt==2.8.0
cryptography==44.0.1

# Cache Redis pour production
django-redis==5.4.0
redis==5.0.1

# Utilitaires pour RAWG API
python-slugify==8.0.1